Add auto-detection for context managers, so they're not detected as iterators
In 0.9.0, sphinxcontrib-trio's sniffing logic will detect a function
like this:

```python
@contextmanager
def closing(obj):
    try:
        yield obj
    finally:
        obj.close()
```

as being *both* a context manager *and* a generator. Because, well, it
is. And we can tell because we walk the `__wrapped__` chain, which we
have to do to detect things like async classmethods. But here the
generator part is an internal detail, invisible to the user, so we
don't want to make that visible. (And a similar problem happens in the
trio docs for functions using our `@acontextmanager` decorator.)

So this commit:

* Adds heuristic sniffing logic for context managers and async context
  managers (specifically it can detect contextlib, contextlib2, and
  any cooperating libraries; trio itself will start cooperating
  shortly).

* Adds a heuristic that any given function cannot be more than one of
  async, an (async) context manager, or an (async) generator.

Technically this will give the wrong result for

```python
async def foo():
    return await create_something()
foo.__returns_contextmanager__ = True

```

but (a) this API is really weird and probably buggy (you should
acquire resources inside the `__enter__` method, not the class
constructor, to make sure they're released!), and (b) sorry, it's a
heuristic. You can always override it manually if you really want to.